fix: decimals type

+++ contracts/shared/Structs.sol
@@ -28,13 +28,13 @@ contract Structs {
         string name;
         string symbol;
         uint256 totalSupply;
-        uint256 decimals;
+        uint8 decimals;
         address issuer;
     }
 
     struct ERC20AssetCommonState {
         address contractAddress;
-        uint256 decimals;
+        uint8 decimals;
         string name;
         string symbol;
         AssetCommonState commonState;
@@ -190,7 +190,7 @@ contract Structs {
         string name;
         string symbol;
         uint256 totalSupply;
-        uint256 decimals;
+        uint8 decimals;
         address issuer;
         bool assetApprovedByIssuer;
         uint256 totalAmountRaised;
